requirements.txt
Django==1.10.5
Pillow==3.2.0
django-allauth==0.25.2
django-bootstrap3==7.0.1
django-decorator-include==1.1
django-multiselectfield==0.1.4
django-phonenumber-field==1.1.0
django-recaptcha==1.0.5
django-rest-multiple-models==1.8.0
django-settings-export==1.1.0
django-stronghold==0.2.7
djangorestframework==3.3.3
geopy==1.11.0
django-timezone-field==2.0
timezonefinder==1.5.7
